Fix remote_form_for to work with Rails 5
remote_form_for is doing almost the same job as FormHelper#form_for from
Rails 5 actionview gem. However, the code had no chance of working in
the original form, because it was handling its list of arguments in the
old convention which was changed sometime in Rails 4 (sic!).

The fix is to copy some code from FormHelper#form_for so that
FormHelper#fields_for gets called with correct arguments.